Add logging decorator to campaign service
This commit adds a new file called logging_decorator.go to the campaign package. The logging_decorator.go file contains a new struct called LoggingServiceDecorator, which is a decorator for the campaign service. The decorator adds logging functionality to the CreateCampaign, UpdateCampaign, GetCampaignByID, GetAllCampaigns, DeleteCampaign, FetchCampaign, and ComposeEmail methods of the campaign service. The decorator logs various information such as the name of the campaign being created, the ID of the campaign being updated, and the ID of the campaign being fetched or deleted. The decorator also logs any errors that occur during these operations.

The purpose of this change is to add logging to the campaign service in order to improve observability and troubleshooting capabilities.